Suddenly this relatively simple change also speeds up some cases

before 
```
Benchmark                            Mode  Cnt     Score    Error   Units
JmhTest._bothifyExpression          thrpt    5  4277.270 ± 41.760  ops/ms
JmhTest._csv                        thrpt    5    35.568 ±  0.919  ops/ms
JmhTest._letterifyExpression        thrpt    5  4629.480 ± 62.134  ops/ms
JmhTest._optionsExpression          thrpt    5  4846.367 ± 34.680  ops/ms
JmhTest._regexifyExpression         thrpt    5  2070.759 ± 23.280  ops/ms
JmhTest.oneMethodCallExpression     thrpt    5  2309.779 ± 26.963  ops/ms
JmhTest.threeMethodsCallExpression  thrpt    5   565.426 ±  5.828  ops/ms
```

after
```

Benchmark                            Mode  Cnt     Score    Error   Units
JmhTest._bothifyExpression          thrpt    5  4538.392 ± 60.038  ops/ms
JmhTest._csv                        thrpt    5    37.199 ±  0.434  ops/ms
JmhTest._letterifyExpression        thrpt    5  4834.415 ± 89.559  ops/ms
JmhTest._optionsExpression          thrpt    5  4894.174 ± 59.062  ops/ms
JmhTest._regexifyExpression         thrpt    5  2099.060 ± 35.080  ops/ms
JmhTest.oneMethodCallExpression     thrpt    5  4688.254 ± 71.117  ops/ms
JmhTest.threeMethodsCallExpression  thrpt    5   590.847 ±  7.354  ops/ms
```